requirements1.txt
#python==3.10.11
streamlit==1.38.0
langchain-community==0.2.16
langchain-text-splitters==0.2.4
langchain-chroma==0.1.3
langchain-huggingface==0.0.3
langchain-groq==0.1.9
unstructured==0.15.0
unstructured[pdf]==0.15.0
nltk==3.8.1

#pip install langchain-huggingface langchain_community
